Add support for subscriptions
This also brings the Redis protocol a bit more up-to-date and no longer
treats commands differently (using multi-bulk for everything).

I also rewrote the command parser as subscriptions need more
flexibility; however it now also avoids using callbacks quite as often
so should be faster in the general case when all the data is in the read
buffer.