added failing lrange test
I am getting errors on some lrange commands with large responses.

lrange.t demonstrates the issue. After the AE::handle's read buffer
is empty it attempts to read more data but does not get any more.
It ends up in an endless loop trying to read the rest of the list.